Rearrange types in "sdk/tokens" package
