+++ packages/sdk/tokens/methods/collection-by-id/types.ts
@@ -1,9 +1,5 @@
 import { CollectionInfoWithOldProperties } from '../create-collection-ex/types';
 
-export interface CollectionIdArguments {
-  collectionId: number;
-}
-
 export interface CollectionInfo extends CollectionInfoWithOldProperties {
   id: number;
   owner: string;
