+++ packages/sdk/tokens/types/index.ts
@@ -1,5 +1,3 @@
-import { CollectionIdArguments } from '../methods/collection-by-id/types';
-
 export * from '../methods/create-collection-ex/types';
 export * from '../methods/create-collection-ex-new/types';
 export * from '../methods/collection-by-id/types';
@@ -22,35 +20,4 @@ export * from '../methods/token-properties/types';
 export * from '../methods/property-permissions/types';
 export * from '../methods/token-by-id/types';
 
-export interface TokenIdArguments extends CollectionIdArguments {
-  tokenId: number;
-}
-
-export interface BurnTokenArguments extends TokenIdArguments {
-  address: string;
-}
-export interface TransferTokenArguments extends TokenIdArguments {
-  from: string;
-  to: string;
-}
-
-export type CollectionProperty = {
-  key: string;
-  value: string;
-};
-
-export type TokenProperty = {
-  key: string;
-  value: string;
-};
-
-export type PropertyPermission = {
-  mutable: boolean;
-  collectionAdmin: boolean;
-  tokenOwner: boolean;
-};
-
-export type PropertyKeyPermission = {
-  key: string;
-  permission: PropertyPermission;
-};
+export * from './shared';
