+++ packages/sdk/tokens/types/shared.ts
@@ -1,0 +1,36 @@
+export interface CollectionIdArguments {
+  collectionId: number;
+}
+
+export interface TokenIdArguments extends CollectionIdArguments {
+  tokenId: number;
+}
+
+export interface BurnTokenArguments extends TokenIdArguments {
+  address: string;
+}
+export interface TransferTokenArguments extends TokenIdArguments {
+  from: string;
+  to: string;
+}
+
+export type CollectionProperty = {
+  key: string;
+  value: string;
+};
+
+export type TokenProperty = {
+  key: string;
+  value: string;
+};
+
+export type PropertyPermission = {
+  mutable: boolean;
+  collectionAdmin: boolean;
+  tokenOwner: boolean;
+};
+
+export type PropertyKeyPermission = {
+  key: string;
+  permission: PropertyPermission;
+};
